Add Identity

+++ Jorje.TheWorld.Domain/Stop.cs
@@ -8,8 +8,8 @@ namespace Jorje.TheWorld.Domain
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         public int Order { get; set; }
         public DateTime Arrival { get; set; }
     }
